--HG--
extra : convert_revision : ea62eb0f88d3efd1b661fe28edbc1d6c964cc05d

+++ libpurple/upnp.c
@@ -352,7 +352,7 @@ gaim_upnp_parse_description_response(const gchar* httpResponse, gsize len,
 			start = start ? start + 3 : baseURL;
 			path = strchr(start, '/');
 			length = path ? path - baseURL : strlen(baseURL);
-			controlURL = g_strdup_printf("%.*s%s", length, baseURL, tmp);
+			controlURL = g_strdup_printf("%.*s%s", (int)length, baseURL, tmp);
 		} else {
 			controlURL = g_strdup_printf("%s%s", baseURL, tmp);
 		}
